added router links and restructured the app to have elements outside the router component
