--- pages/dashboard.js
@@ -1,9 +1,0 @@
-import Container from 'react-bootstrap/Container'
-
-export default function Dashboard() {
-	return (
-		<Container className='my-4'>
-			<h1>Welcome to the F1 Hub</h1>
-		</Container>
-	)
-}
